pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "crc-service-portal"
version = "0.0.0"
readme = "README.md"
description = "Self service portal for the Pitt CRC."
authors = ["Pitt Center for Research Computing"]
keywords = ["Pitt", "CRC", "HPC", "django"]

[tool.poetry.scripts]
crc-service-portal = "crc_service_portal.manage:main"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "1.0.0"
